+++ packages/planets/src/quadtree/Quadtree.ts
@@ -1,13 +1,12 @@
 import * as THREE from "three"
 import { Vector3 } from "three"
 
-const CHILD_SIZE_X_COMPARATOR = 1.25
-
 export interface QuadTreeParams {
   localToWorld: THREE.Matrix4
   size: number
   minNodeSize: number
   origin: THREE.Vector3
+  comparatorValue: number
 }
 
 export interface Node {
@@ -28,6 +27,9 @@ export class QuadTree {
       new THREE.Vector3(-s, -s, 0),
       new THREE.Vector3(s, s, 0),
     )
+    if (params.comparatorValue <= 0) {
+      throw new Error("Quadtree Comparison Value must be greater than 0");
+    }
     this.origin = params.origin
     this.root = {
       bounds: b,
@@ -69,7 +71,7 @@ export class QuadTree {
     const distToChild = this.#distanceToChild(child, pos)
 
     if (
-      distToChild < child.size.x * CHILD_SIZE_X_COMPARATOR &&
+      distToChild < child.size.x * this.params.comparatorValue &&
       child.size.x > this.params.minNodeSize
     ) {
       child.children = this.#createChildren(child)