+++ packages/planets/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hello-worlds/planets
 
+## 0.0.13
+
+### Patch Changes
+
+- expose lodDistanceComparisonValue to allow users to tune the resolution & performance
+
 ## 0.0.12
 
 ### Patch Changes